pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "osm-powerplants"
version = "0.1.0"
description = "Extract and process power plant data from OpenStreetMap"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "Open Energy Transition", email = "info@openenergytransition.org" },
]
keywords = ["openstreetmap", "power plants", "energy", "osm", "overpass"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering",
]
dependencies = [
    "pandas>=2.0",
    "numpy>=1.24",
    "requests>=2.28",
    "pycountry>=22.3",
    "shapely>=2.0",
    "diskcache>=5.6",
    "tqdm>=4.65",
    "scikit-learn>=1.2",
    "python-dateutil>=2.8",
    "platformdirs>=3.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "ruff>=0.8",
    "pre-commit>=3.5",
]
docs = [
    "mkdocs>=1.5",
    "mkdocs-material>=9.4",
    "mkdocstrings[python]>=0.24",
]

[project.scripts]
osm-powerplants = "osm_powerplants.cli:main"

[project.urls]
Homepage = "https://github.com/open-energy-transition/osm-powerplants"
Repository = "https://github.com/open-energy-transition/osm-powerplants"
Documentation = "https://open-energy-transition.github.io/osm-powerplants"

[tool.hatch.build.targets.wheel]
packages = ["src/osm_powerplants"]

[tool.hatch.build.targets.wheel.force-include]
"src/osm_powerplants/config.yaml" = "osm_powerplants/config.yaml"

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
]
